cambio serv

+++ src/main/java/com/guardias/backend/repositorio/RegistroActividadRepositorio.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.guardias.backend.modelo.RegistroActividad;
 
 @Repository
-public interface RegistroActividadRepositorio extends JpaRepository<RegistroActividad,Long>{
+public interface RegistroActividadRepositorio extends JpaRepository<RegistroActividad,Integer>{
 
     /* Optional<RegistroActividad> findById(int id); */
     boolean existsById(int id);
